Crate mentat_transaction

Re-exports

pub use entity_builder::InProgressBuilder;

pub use entity_builder::TermBuilder;

pub use metadata::Metadata;

Modules

entity_builder

metadata

query

Structs

InProgress
Represents an in-progress, not yet committed, set of changes to the store. Call commit to commit your changes, or rollback to discard them. A transaction is held open until you do so. Your changes will be implicitly dropped along with this struct.

InProgressRead
Represents an in-progress set of reads to the store. Just like InProgress, which is read-write, but only allows for reads.

Enums

CacheAction

CacheDirection

Traits

Pullable

Queryable